After initial switching setup, switch DHCP to external iface
After the external interface has been bound to Open vSwitch, we should
not attempt to DHCP on it, as it won't receive the DHCP responses.
Attempting to DHCP results in startup stalling on network-online.target,
because systemd-networkd-wait-online.service looks at every network
interface it should configure based on the [Match] sections, and waits
for all of them to be configured by systemd-networkd, which will never
finish if DHCP is broken.