Rejigger networking so the bridge has the MAC of the external interface
This allows the subsequent DHCP request to get the same IP address back,
which means it only needs one address for first boot, and the address
can be pre-allocated by the DHCP server before deployment.

This is needed for the floating address range to be allocated in some
set-ups, such as our local one at the office, where a subnet in a
different class is routed to the server.