openstack: Disable nova firewall management
This should be handled by neutron, and except for the mis-configuration,
it should have been.

However, since both neutron and nova were configured to handle
firewalling, they would both install their firewall rules into iptables,
and it would be random which one would be used as either service is
likely to start before the other and install their hook first.

The result being that we'd randomly not be able to reach VMs after a
reboot, unless we'd installed the same firewall rules in both nova and
neutron.