Mark that we use definitions version 3 in this repo.
Commit 7b23e26d1e2bb24ca72a186 added a configure extension to all
systems (install-essential-files.configure). From that point on, the
only build tool able to deploy these definitions in Morph
commit ed423dc974a61f1c033491 and newer.

The first tagged version of Morph that contains this commit is
'baserock-definitions-v3'. Thus, we should consider the
install-essential-files.configure extension part of definitions format
version 3.

It was already impossible to build the 15.19 release of Baserock with
the version of Morph from the 15.10 release, due to other incompatible
changes that were made to the definitions since 15.10 was released.

Change-Id: I05c67cd039bb2124fef667bfe38a84dd9887b646