Fix 'name' field mismatches
The 'name' field is set in both a chunk's definition, and in the
chunk reference point in each stratum where that chunk is included.
Some of them didn't agree on the name of the chunk. In all cases I've
made 'name' match the filename of the chunk .morph file, either by
fixing the chunk .morph file or the stratum .morph file (and in one
case, the filename).

This should have no effect on behaviour, it's just a tidyup.

Change-Id: I90bcd3fa382e24c8aa2d370ea5f545491169d1db