Reformat all definitions according to a consistent style
This was done using the 'indent' tool, which uses a fork of PyYAML named
'ruamel.yaml' to rewrite YAML files without losing comments, ordering,
or certain elements of formatting.

My aim with doing this is to open the door to automated editing of the
reference system definitions using the 'ruamel.yaml' library. This can
be used to implement automated migration when we want to make changes to
the YAML format that we use to represent Baserock system definitions.

Although this looks drastic, remember that it's actually only altered
65 out of 608 .morph files -- the vast majority already pass unchanged
through my version of ruamel.yaml.


Change-Id: I95ec978714b5bd1c02c90183336a9fbb846cb692