Add ruamel.yaml library to 'build' and 'devel' systems
This is currently done by adding it to the 'morph-utils' stratum,
because it is used as a tool for working with .morph files.

The ruamel.yaml library allows writing YAML files while preserving
comments, ordering and some formatting form the input files. This makes
it more practical for us to write automated migration scripts for
Baserock definitions.

Change-Id: I886232dad4d64bcf1803f52e94ae25b0fa2d9969