openstack: fix the public endpoints
The public keystone endpoints ought to be accessible outside of the
three-node cluster.  However, the CONTROLLER_HOST_ADDRESS parameter
corresponds to an hostname that maps to an IP addess in the private
management network.  Resolve this by creating the public endpoints with
the controller hostname, which should be public accessible.

In addition, the admin endpoint for keystone needs to be public
accessible as well.

Also, remove unneeded entries in the hosts files.  E.g
threenode-controller could resolve to a private or public IP address,
depending on the configuration on nsswitch.conf.

Change-Id: Ibb305d0541ee4094fcc0aa1f8233a5d98fed58e5