install-files: Allow definition of manifests in multiple variables
install-files was initially designed such that if you wanted to install
multiple manifests you should put multiple whitespace separated paths in
the INSTALL_FILES variable.

This is clunky if you want to extend the set of files to install,
particularly in cluster deployments with multiple systems,
as otherwise you could make use of YAML merge keys.

So now, like fstab.configure you can set multple options, by using
environment variables that start with INSTALL_FILES.

It's also easier to extend the set of install files on the command-line
or in the environment this way too, since you don't need to parse the
YAML to see what it originally had.

Change-Id: I298db7ae9790a536e5a5f32b069d0908012695af