Migrate to definitions format version 7: add a DEFAULTS file
This is the DEFAULTS file embedded in
migrations/007-defaults-in-definitions.py, which is exactly the settings
that are also built into Morph.

I expect the DEFAULTS file used in definitions.git to change over time,
but we don't need to keep the one in
migrations/007-defaults-in-definitions.py in sync. People who have
forked definitions and want to keep their DEFAULTS in line with ours
should use `git merge`.

So it might seem like migrations/007-defaults-in-definitions script is
not useful (`git merge` could have handled adding DEFAULTS, too),
but we need to have a migration for *every* version so that you can
easily go from V1 to V10, or whatever, without having to do a 'git
merge' from definitions.git first.

Change-Id: Idf19c31da013eddcf2e1f165086e697ca1805498