Explicitly disable guile in build of gnu make.
During the stage2 build, we are using header files from the target
sandbox but host build tools, which will automatically detect guile
if present on the build host.

This prevents the build of 'make' to otherwise fail while building
guile.o, due to the lack of 'gmp.h' (and possibly other unneeded cruft).

Change-Id: I5be8467afc9486bf40055000e4c6460232345b93