RFC: Build git with man pages
This commit renames git in core to git-minimal and adds a git chunk into
devtools, the git chunk will replace git-minimal and will build and
install git's man pages.

It seems difficult to modify git in core so that it builds with docs
without creating circular dependencies, chunks such as libtool depend
on git, libxml2 depends on libtool, so libxml2 depends on git, but
docutils depends on libxml2, and we need to make git depend on docutils,
the fact that python2 is no longer in core also adds to the difficulty
since asciidoc's build process depends on python2.

Change-Id: Ifc64019b941740c85db058d36183b4a210c60e15