mason: clean-up all the cache after every run
This way we never fill up the disk. We don't care about the cache
given that the distbiuld network uploads everything to the artifacts server.

The solution is not very clever, but will help us to avoid Mason going
red just because of disk space. Also, I don't think Mason deserves more
attention.

Change-Id: Ifbdaac0e992fe9cc4736010fd26dc9c9e7beb7f4