Add xdg-app-common stratum
xdg-app is a tool that uses OSTree to install and update runtimes and applications.

It also allows the user to start these applications in a container-like environment
where they are isolated (to various degrees) from the host operating system.
In particular, each application depends on a particular runtime which
supplies /usr for the application.

Xdg-App also has commands related to building and distributing application bundles.

https://wiki.gnome.org/Projects/SandboxedApps

Change-Id: Ib5b01158b032d52b77b19fc67aa23577c357154e