Adding gnome-initial-setup and dependencies to gnome stratum
This change adds the following chunks to the gnome stratum:
  o cracklib
  o libpwquality (uses cracklib to provide password strength checking)
  o librest      (RESTful API library, needed by online accounts)
  o krb5         (Kerberos Network Authentication)
  o gnome-online-accounts
  o gnome-initial-setup

Additionally, evolution-data-server is now compiled with online account
support - all of the above dependencies are hard requirements for
the gnome initial setup tool.

The WebKitGtk stratum is now a dependency of GNOME and is also
added to the GNOME system with this commit (again required by
the gnome-online-accounts module).

Change-Id: I103402bc9125b1e9ca57b393b1cef6fb99f526e3