Proper configuration of pulseaudio
Pulseaudio normally runs as a user specific service and
not a system service, some warnings were popping up because
we had passed /lib/systemd/system as --systemduserunitdir
(note that that is the *user* unit directory).

Instead we now pass /usr/lib/systemd/user as the
appropriate --systemduserunitdir which causes the .service
and .socket files to be installed in the appropriate
location, and systemd activates pulseaudio automatically
via regular socket activation.

Also, it is no longer necessary to use custom install commands
creating a symlink for the service, as the service need not
be "enabled", it is just activated upon request.

Change-Id: Ic1a246643c45062a4432be3f739de9a7146e5e5e