Fix minimal systems build
This was broken by changes in c4bf91813c787b503eaeaa1d372f271d4638a16d
which expected 'depmod' to be present in all systems.

It would be annoying to prevent anyone from using modules, but it also
defeats the point of minimal-system to start adding bits of the
'foundation' stratum.

This hack to only run 'depmod' if it is present in the system seems like
a reasonable compromise, until we have a nice way of customising the
Linux config for different systems.

Change-Id: Ida083b51353cab78f8d3f6a72af711a00d0fcfd8