cloud-init.configure: Fix use of MORPH_LOG_FD
MORPH_LOG_FD exists so that deployment extensions can write detailed log
info to a file. This extension doesn't output much detailed log info,
but some of the Python ones do & it's really useful for debugging.

With Morph, this extension seems to have been writing logs to a file
named '3' rather than to the actual deployment log.

With YBD, this extension crashed because MORPH_LOG_FD was unset. Extra logs
will now go to stdout when deploying with YBD.

Change-Id: I46ead092d0e071970e355c8539823836f28dd97b