parted.morph: Remove the perl requirement from bootstrap.conf
The gnulib bootstrap script is idiotic, when it does get_version(),
it looks for a string v${major}.${minor}.${micro}... it starts this
by taking the last 'v' on the line and removing everything else.

On an arm arch, instead of getting 5.22.0 from the (v5.22.0) string
output from perl --version... it gets 5tejl-linux from the armv5tejl-linux
string on the same line... get_version() thus thinks we have "perl 5" and
breaks.

This is not a problem for armv7lhf builds, because the bootstrap script
conveniently interprets perl's version as perl version 7, which is newer
than the required 5.6.

Change-Id: I1445cfd84efd37c1828eae09def88b27f18af250